requirements.txt
apex==0.9.10dev
click==8.0.3
easydict==1.9
fitlog==0.9.13
nltk==3.6.7
nni==2.5
numpy==1.22.0
opt_einsum==3.3.0
prettytable==2.5.0
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
setuptools==58.0.4
spacy==3.2.4
tensorboardX==2.5
torch==1.10.0
torchtext==0.4.0
tqdm==4.62.3
